requirements.txt
# Discord Bot Dependencies
discord.py>=2.3.0
python-dotenv>=1.0.0

# Database Dependencies
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
alembic>=1.12.0

# HTTP Requests
aiohttp>=3.8.0

# RCON (опционально, если используется RCON функционал)
# gamercon-async>=0.1.0

